requirements.txt
# YOLOv5 Flask API - Final Requirements
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# YOLOv5 및 필수 라이브러리 (PyTorch는 Dockerfile에서 설치)
ultralytics>=8.0.0
pillow>=9.0.0
opencv-python-headless>=4.5.0
numpy<2.0.0
pyyaml>=6.0
requests>=2.28.0
psutil>=5.9.0
